Move write_chart_data_to_file to create_graph.py
